tests/wpt/web-platform-tests/loading/lazyload/invisible-lazy-image.tentative.html

<!DOCTYPE html>
<head>
  <title>Test behavior of in viewport invisible lazy images</title>
  <link rel="author" title="Rob Buis" href="mailto:rbuis@igalia.com">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="common.js"></script>
</head>

<body>
  <!-- These two images should load, the latter not blocking the window load event -->
  <img id="expected" src='resources/image.png?expected&pipe=trickle(d1)'>
  <img id="visibility_hidden" style="visibility:hidden;"
       src='resources/image.png?visibility_hidden&pipe=trickle(d2)' loading='lazy'
       onload="visibility_hidden_img.resolve();" onerror="visibility_hidden_img.reject();">

  <!-- These images should not load at all -->
  <img id="display_none" style="display:none;" src='resources/image.png?display_none'
       loading='lazy'
       onload="display_none_img.resolve();" onerror="display_none_img.reject();">
  <img id="attribute_hidden" hidden src='resources/image.png?attribute_hidden' loading='lazy'
       onload="attribute_hidden_img.resolve();" onerror="attribute_hidden_img.reject();">
  <img id="js_display_none" src='resources/image.png?js_display_none' loading='lazy'
       onload="js_display_none_img.resolve();" onerror="js_display_none_img.reject();">
  <script>
    document.getElementById("js_display_none").style = 'display:none;';
  </script>
</body>

<!--
Marked as tentative until https://github.com/whatwg/html/pull/3752 is landed.
-->

<script>
  const expected = document.getElementById("expected");
  const visibility_hidden_img = new ElementLoadPromise("visibility_hidden");
  const display_none_img = new ElementLoadPromise("display_none");
  const attribute_hidden_img = new ElementLoadPromise("attribute_hidden");
  const js_display_none_img = new ElementLoadPromise("js_display_none");

  let has_window_loaded = false;

  async_test(t => {
    window.addEventListener("load", t.step_func(() => {
      has_window_loaded = true;
    }));

    const unreached_not_rendered_img_func =
      t.unreached_func("The not-rendered in-viewport loading=lazy images " +
                       "should not have attempted to load.");

    display_none_img.promise
      .then(unreached_not_rendered_img_func)
      .catch(unreached_not_rendered_img_func);

    attribute_hidden_img.promise
      .then(unreached_not_rendered_img_func)
      .catch(unreached_not_rendered_img_func);

    js_display_none_img.promise
      .then(unreached_not_rendered_img_func)
      .catch(unreached_not_rendered_img_func);

    visibility_hidden_img.promise.then(
      t.step_func_done(() => {
        assert_true(is_image_fully_loaded(visibility_hidden_img.element(), expected),
                    "The loading=lazy visibility:hidden image is equivalent " +
                    "to the expected image.");
        assert_true(has_window_loaded,
                    "The loading=lazy visibility:hidden image does not block " +
                    "the window load event, and finishes loading after the " +
                    "window load event fires.");
      })
    ).catch(t.unreached_func("The loading=lazy visibility:hidden image " +
                             "should load successfully."));
  }, "Test behavior of in viewport invisible lazy images");
</script>
