third_party/blink/web_tests/external/wpt/loading/lazyload/invisible-lazy-image.tentative-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: Test behavior of in viewport invisible lazy images
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
FAIL Test behavior of in viewport invisible lazy images assert_true: The loading=lazy visibility:hidden image does not block the window load event, and finishes loading after the window load event fires. expected true got false
Harness: the test ran to completion.
